XD-796: Split shell and job 'all' commands
Replaces 'all' option to stream and job 'deploy/undeploy'
and 'destroy' commands with new 'stream all' or 'job all'
syntax.

+++ spring-xd-shell/src/main/java/org/springframework/xd/shell/command/JobCommands.java
@@ -26,7 +26,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.xd.rest.client.JobOperations;
 import org.springframework.xd.rest.client.domain.JobDefinitionResource;
 import org.springframework.xd.shell.XDShell;
-import org.springframework.xd.shell.util.Assertions;
 import org.springframework.xd.shell.util.Table;
 import org.springframework.xd.shell.util.TableHeader;
 import org.springframework.xd.shell.util.TableRow;
@@ -49,11 +48,20 @@ public class JobCommands implements CommandMarker {
 
 	private final static String DEPLOY_JOB = "job deploy";
 
+	private final static String DEPLOY_ALL_JOBS = "job all deploy";
+
 	private final static String LAUNCH_JOB = "job launch";
 
 	private final static String UNDEPLOY_JOB = "job undeploy";
 
+	private final static String UNDEPLOY_ALL_JOBS = "job all undeploy";
+
 	private final static String DESTROY_JOB = "job destroy";
+
+	private final static String DESTROY_ALL_JOBS = "job all destroy";
+
+	@Autowired
+	private UserInput userInput;
 
 	@Autowired
 	private XDShell xdShell;
@@ -88,27 +96,27 @@ public class JobCommands implements CommandMarker {
 		return table;
 	}
 
-	@CliCommand(value = DEPLOY_JOB, help = "Deploy previously created job(s)")
+	@CliCommand(value = DEPLOY_JOB, help = "Deploy a previously created job")
 	public String deployJob(
-			@CliOption(key = { "", "name" }, help = "the name of the job to deploy", optionContext = "existing-job disable-string-converter") String name,
-			@CliOption(key = { "all" }, help = "deploy all the existing jobs", unspecifiedDefaultValue = "false", specifiedDefaultValue = "true") boolean all,
+			@CliOption(key = { "", "name" }, help = "the name of the job to deploy", mandatory = true, optionContext = "existing-job disable-string-converter") String name,
 			@CliOption(key = "dateFormat", help = "the optional date format for job parameters") String dateFormat,
 			@CliOption(key = "numberFormat", help = "the optional number format for job parameters") String numberFormat,
 			@CliOption(key = "makeUnique", help = "shall job parameters be made unique?") Boolean makeUnique) {
-		String message = "";
-		switch (Assertions.exactlyOneOf("name", name, "all", all)) {
-			case 0:
-				jobOperations().deployJob(name, dateFormat, numberFormat, makeUnique);
-				message = String.format("Deployed job '%s'", name);
-				break;
-			case 1:
-				jobOperations().deployAll();
-				message = String.format("Deployed all the jobs");
-				break;
-			default:
-				throw new IllegalArgumentException("You must specify exactly one of 'name', 'all'");
+		jobOperations().deployJob(name, dateFormat, numberFormat, makeUnique);
+		return String.format("Deployed job '%s'", name);
+	}
+
+	@CliCommand(value = DEPLOY_ALL_JOBS, help = "Deploy previously created job(s)")
+	public String deployAllJobs(
+			@CliOption(key = "force", help = "bypass confirmation prompt", unspecifiedDefaultValue = "false", specifiedDefaultValue = "true") boolean force
+			) {
+		if (force || "y".equalsIgnoreCase(userInput.prompt("Really deploy all jobs?", "n", "y", "n"))) {
+			jobOperations().deployAll();
+			return String.format("Deployed all jobs");
 		}
-		return message;
+		else {
+			return "";
+		}
 	}
 
 	@CliCommand(value = LAUNCH_JOB, help = "Launch previously deployed job")
@@ -119,44 +127,47 @@ public class JobCommands implements CommandMarker {
 		return String.format("Successfully launched the job '%s'", name);
 	}
 
-	@CliCommand(value = UNDEPLOY_JOB, help = "Un-deploy existing job(s)")
+	@CliCommand(value = UNDEPLOY_JOB, help = "Un-deploy an existing job")
 	public String undeployJob(
-			@CliOption(key = { "", "name" }, help = "the name of the job to un-deploy", optionContext = "existing-job disable-string-converter") String name,
-			@CliOption(key = { "all" }, help = "undeploy all the existing jobs", unspecifiedDefaultValue = "false", specifiedDefaultValue = "true") boolean all) {
-		String message = "";
-		switch (Assertions.exactlyOneOf("name", name, "all", all)) {
-			case 0:
-				jobOperations().undeploy(name);
-				message = String.format("Un-deployed Job '%s'", name);
-				break;
-			case 1:
-				jobOperations().undeployAll();
-				message = String.format("Un-deployed all the jobs");
-				break;
-			default:
-				throw new IllegalArgumentException("You must specify exactly one of 'name', 'all'");
-		}
-		return message;
+			@CliOption(key = { "", "name" }, help = "the name of the job to un-deploy", mandatory = true, optionContext = "existing-job disable-string-converter") String name
+			) {
+		jobOperations().undeploy(name);
+		return String.format("Un-deployed Job '%s'", name);
 	}
 
-	@CliCommand(value = DESTROY_JOB, help = "Destroy existing job(s)")
-	public String destroyJob(
-			@CliOption(key = { "", "name" }, help = "the name of the job to destroy", optionContext = "existing-job disable-string-converter") String name,
-			@CliOption(key = { "all" }, help = "destroy all the existing jobs", unspecifiedDefaultValue = "false", specifiedDefaultValue = "true") boolean all) {
-		String message = "";
-		switch (Assertions.exactlyOneOf("name", name, "all", all)) {
-			case 0:
-				jobOperations().destroy(name);
-				message = String.format("Destroyed job '%s'", name);
-				break;
-			case 1:
-				jobOperations().destroyAll();
-				message = String.format("Destroyed all the jobs");
-				break;
-			default:
-				throw new IllegalArgumentException("You must specify exactly one of 'name', 'all'");
+	@CliCommand(value = UNDEPLOY_ALL_JOBS, help = "Un-deploy all existing jobs")
+	public String undeployAllJobs(
+			@CliOption(key = "force", help = "bypass confirmation prompt", unspecifiedDefaultValue = "false", specifiedDefaultValue = "true") boolean force
+			) {
+		if (force || "y".equalsIgnoreCase(userInput.prompt("Really undeploy all jobs?", "n", "y", "n"))) {
+			jobOperations().undeployAll();
+			return String.format("Un-deployed all the jobs");
 		}
-		return message;
+		else {
+			return "";
+		}
+	}
+
+	@CliCommand(value = DESTROY_JOB, help = "Destroy an existing job")
+	public String destroyJob(
+			@CliOption(key = { "", "name" }, help = "the name of the job to destroy", mandatory = true, optionContext = "existing-job disable-string-converter") String name
+			) {
+		jobOperations().destroy(name);
+		return String.format("Destroyed job '%s'", name);
+	}
+
+	@CliCommand(value = DESTROY_ALL_JOBS, help = "Destroy all existing jobs")
+	public String destroyAllJobs(
+			@CliOption(key = "force", help = "bypass confirmation prompt", unspecifiedDefaultValue = "false", specifiedDefaultValue = "true") boolean force
+
+			) {
+		if (force || "y".equalsIgnoreCase(userInput.prompt("Really destroy all jobs?", "n", "y", "n"))) {
+			jobOperations().destroyAll();
+			return String.format("Destroyed all the jobs");
+		}
+		else {
+			return "";
+		}
 	}
 
 	private JobOperations jobOperations() {
